+++ src/main/java/odyssey/backend/domain/roadmap/exception/RoadmapNotFoundException.java
@@ -1,5 +1,6 @@
 package odyssey.backend.domain.roadmap.exception;
 
+import odyssey.backend.domain.roadmap.exception.error.RoadmapExceptionProperty;
 import odyssey.backend.shared.exception.GlobalException;
 
 public class RoadmapNotFoundException extends GlobalException {
